requirements_ubuntu18.04.txt
# pip3 install -r requirements.txt

# base ----------------------------------------
imgviz>=0.11.0
matplotlib
numpy
Pillow>=2.8.0
PyQt5
qtpy
PyYAML
termcolor
lxml
imagecodecs
colordict

# coco ----------------------------------------
pycocotools>=2.0